requirements.txt
paho-mqtt
simplejson
unicornhat
